refactor: Subscription.Amount を decimal から int に変更 (#21)
日本円は小数点以下を持たないため、型レベルで不正値を防ぐ目的で
Subscription.Amount を decimal から int に変更する。
Transaction.Amount（#5）との型の一貫性も確保する。

- Domain/Entities/Subscription.cs: Amount を decimal → int に変更
- Infrastructure/Persistence/Configurations/SubscriptionConfiguration.cs:
  HasColumnType("numeric(12,0)") → "integer" に変更
- Migration: ChangeSubscriptionAmountToInteger を追加
  (subscriptions.amount: numeric(12,0) → integer)
- tests: decimal リテラル (1490m) を int (1490) に修正、不要なキャストを削除
- docs/db-design.md: ER図・テーブル定義・設計判断セクションを更新

+++ backend/src/KakeiBase.WebApi/Domain/Entities/Subscription.cs
@@ -6,14 +6,14 @@ public class Subscription
     public Guid UserId { get; private set; }
     public Guid CategoryId { get; private set; }
     public string Name { get; private set; } = string.Empty;
-    public decimal Amount { get; private set; }
+    public int Amount { get; private set; }
     public bool IsActive { get; private set; }
     public DateTimeOffset CreatedAt { get; private set; }
     public DateTimeOffset UpdatedAt { get; private set; }
 
     private Subscription() { }
 
-    public static Subscription Create(Guid userId, Guid categoryId, string name, decimal amount)
+    public static Subscription Create(Guid userId, Guid categoryId, string name, int amount)
     {
         var now = DateTimeOffset.UtcNow;
         return new Subscription
